eclypso/srv/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.6
pydantic==2.5.2
pydicom==2.4.3
nibabel==5.1.0
SimpleITK==2.3.1
pillow==9.4.0
numpy==1.26.4
python-dotenv==1.0.0
httpx==0.25.2 